Move link methods to separate module.
