+++ deepwell/src/methods/page.rs
@@ -189,96 +189,6 @@ pub async fn page_restore(mut req: ApiRequest) -> ApiResponse {
     Ok(body.into())
 }
 
-pub async fn page_links_from_get(req: ApiRequest) -> ApiResponse {
-    let txn = req.database().begin().await?;
-    let ctx = ServiceContext::new(&req, &txn);
-
-    let site_id = req.param("site_id")?.parse()?;
-    let reference = Reference::try_from(&req)?;
-    tide::log::info!("Getting page links for page {reference:?} in site ID {site_id}");
-
-    let page = PageService::get(&ctx, site_id, reference).await.to_api()?;
-    let output = LinkService::get_from(&ctx, page.page_id).await.to_api()?;
-    let body = Body::from_json(&output)?;
-    txn.commit().await?;
-
-    Ok(body.into())
-}
-
-pub async fn page_links_to_get(req: ApiRequest) -> ApiResponse {
-    let txn = req.database().begin().await?;
-    let ctx = ServiceContext::new(&req, &txn);
-
-    let site_id = req.param("site_id")?.parse()?;
-    let reference = Reference::try_from(&req)?;
-    tide::log::info!("Getting page links from page {reference:?} in site ID {site_id}");
-
-    let page = PageService::get(&ctx, site_id, reference).await.to_api()?;
-    let output = LinkService::get_to(&ctx, page.page_id, None)
-        .await
-        .to_api()?;
-
-    let body = Body::from_json(&output)?;
-    txn.commit().await?;
-    Ok(body.into())
-}
-
-pub async fn page_links_to_missing_get(req: ApiRequest) -> ApiResponse {
-    let txn = req.database().begin().await?;
-    let ctx = ServiceContext::new(&req, &txn);
-
-    let site_id = req.param("site_id")?.parse()?;
-    let page_slug = req.param("page_slug")?;
-    tide::log::info!(
-        "Getting missing page links from page slug {page_slug} in site ID {site_id}",
-    );
-
-    let output = LinkService::get_to_missing(&ctx, site_id, page_slug, None)
-        .await
-        .to_api()?;
-
-    let body = Body::from_json(&output)?;
-    txn.commit().await?;
-    Ok(body.into())
-}
-
-pub async fn page_links_external_from(req: ApiRequest) -> ApiResponse {
-    let txn = req.database().begin().await?;
-    let ctx = ServiceContext::new(&req, &txn);
-
-    let site_id = req.param("site_id")?.parse()?;
-    let reference = Reference::try_from(&req)?;
-    tide::log::info!(
-        "Getting external links from page {reference:?} in site ID {site_id}",
-    );
-
-    let page = PageService::get(&ctx, site_id, reference).await.to_api()?;
-    let output = LinkService::get_external_from(&ctx, page.page_id)
-        .await
-        .to_api()?;
-
-    let body = Body::from_json(&output)?;
-    txn.commit().await?;
-    Ok(body.into())
-}
-
-pub async fn page_links_external_to(req: ApiRequest) -> ApiResponse {
-    let txn = req.database().begin().await?;
-    let ctx = ServiceContext::new(&req, &txn);
-
-    let site_id = req.param("site_id")?.parse()?;
-    let url = req.param("url")?;
-    tide::log::info!("Getting external links to URL {url} in site ID {site_id}");
-
-    let output = LinkService::get_external_to(&ctx, site_id, url)
-        .await
-        .to_api()?;
-
-    let body = Body::from_json(&output)?;
-    txn.commit().await?;
-    Ok(body.into())
-}
-
 async fn build_page_response(
     ctx: &ServiceContext<'_>,
     page: &PageModel,
